fixes warning in build, allow compile in macOS environment

+++ src/glyphs.cpp
@@ -572,7 +572,7 @@ void RangeAsync(uv_work_t* req) {
                     // direct type conversions, no need for checking or casting
                     glyph_writer.add_uint32(3, glyph.width);
                     glyph_writer.add_uint32(4, glyph.width);
-                    glyph_writer.add_sint32(5, glyph.width);
+                    glyph_writer.add_sint32(5, static_cast<int32_t>(glyph.width));
 
                     // conversions requiring checks, for safety and correctness
